Move the tilepicker mostly into graphics
